backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
jiwer
requests
openai-whisper
# ffmpeg-static



